Change packages from com.consol.citrus to org.citrusframework
