+++ citrus-remote-maven-plugin/src/main/java/org/citrusframework/remote/plugin/config/ServerConfiguration.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2006-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.citrusframework.remote.plugin.config;
+
+import org.apache.maven.plugins.annotations.Parameter;
+
+import java.io.Serializable;
+
+/**
+ * @author Christoph Deppisch
+ * @since 2.7.4
+ */
+public class ServerConfiguration implements Serializable {
+
+    @Parameter(property = "citrus.remote.server.url", required = true, defaultValue = "http://localhost:8686")
+    private String url = "http://localhost:8686";
+
+    /**
+     * Gets the url.
+     *
+     * @return
+     */
+    public String getUrl() {
+        return url;
+    }
+
+    /**
+     * Sets the url.
+     *
+     * @param url
+     */
+    public void setUrl(String url) {
+        this.url = url;
+    }
+}
